Fix half of style comments. No class TBootstrap yet.

+++ yt/server/skynet_manager/config.h
@@ -79,7 +79,7 @@ class TTombstoneCacheConfig
 {
 public:
     TDuration Ttl;
-    size_t MaxSize;
+    int MaxSize;
 
     TTombstoneCacheConfig()
     {
